backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
# ta-lib==0.4.28  # Optional - requires system TA-Lib library

# Coinbase API
coinbase-advanced-py==1.8.2

# Background tasks
celery==5.3.4
redis==5.0.1

# Real-time communication
python-socketio==5.10.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6

# HTTP client for testing
httpx==0.25.2
